-Rework of the examples architecture -Added a helloworld Ewample project -Corrected bug when neumber of descritor was 0

+++ Library/Source Files/LavaCake/RenderPass.cpp
@@ -92,7 +92,7 @@ namespace LavaCake {
 		}
 
 		void RenderPass::addSubPass(std::vector<GraphicPipeline*> p, uint32_t AttachementFlag, std::vector<uint32_t> input_number) {
-			for (int i = 0; i < p.size(); i++) {
+			for (size_t i = 0; i < p.size(); i++) {
 				p[i]->setSubpassNumber(m_subpass.size());
 			}
 			m_subpass.push_back(p);
